refactor: separate zustand store into 2 stores & remove utils folder, move utils functions to lib folder
