+++ src/components/leftSideBar/playlists/index.tsx
@@ -1,13 +1,12 @@
 import { Link } from 'react-router-dom';
 
-import { useStore } from '@/store';
 import { ScrollArea } from '@/components/ui/scroll-area';
-
 import DeletePlaylistDialog from './delete-playlist-dialog';
 import NewPlaylistPopover from './new-playlist-popover';
+import { useTracksStore } from '@/store/useTracksStore';
 
 function Playlists() {
-  const playlists = useStore(state => state.playlists);
+  const playlists = useTracksStore(state => state.playlists);
 
   return (
     <div>
