--- src/utils/errors.ts
@@ -1,5 +1,0 @@
-export function assertError(error: unknown): asserts error is Error {
-  if (!(error instanceof Error)) {
-    throw error
-  }
-}
